+++ OwnTracks/OwnTracks/Connection.h
@@ -59,4 +59,7 @@ enum state {
 
 - (NSString *)parameters;
 
+- (void)addSubscriptionTo:(NSString *)topic qos:(MQTTQosLevel)qos;
+- (void)removeSubscriptionFrom:(NSString *)topic;
+
 @end