+++ OwnTracks/OwnTracks/LBS.h
@@ -1,0 +1,18 @@
+//
+//  LBS.h
+//  OwnTracks
+//
+//  Created by Christoph Krey on 20.06.15.
+//  Copyright (c) 2015 OwnTracks. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface LBS : NSObject
+@property (strong, nonatomic) NSString *lastGeoHash;
+
+- (void)reset:(NSManagedObjectContext *)context;
+- (void)newLocation:(double)latitude longitude:(double)longitude context:(NSManagedObjectContext *)context;
+- (BOOL)processMessage:(NSString *)topic data:(NSData *)data retained:(BOOL)retained context:(NSManagedObjectContext *)context;
+
+@end