+++ OwnTracks/OwnTracks/MessageTVC.h
@@ -1,0 +1,13 @@
+//
+//  MessageTVC.h
+//  OwnTracks
+//
+//  Created by Christoph Krey on 20.06.15.
+//  Copyright (c) 2015 OwnTracks. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface MessageTVC : UITableViewController <NSFetchedResultsControllerDelegate>
+
+@end