+++ OwnTracks/OwnTracks/coredata/Message.m
@@ -1,0 +1,20 @@
+//
+//  Message.m
+//  OwnTracks
+//
+//  Created by Christoph Krey on 20.06.15.
+//  Copyright (c) 2015 OwnTracks. All rights reserved.
+//
+
+#import "Message.h"
+
+
+@implementation Message
+
+@dynamic topic;
+@dynamic timestamp;
+@dynamic desc;
+@dynamic url;
+@dynamic expiry;
+
+@end