+++ OwnTracks/Pods/Target Support Files/Pods-OwnTracksTests-objc-geohash/Pods-OwnTracksTests-objc-geohash-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Pods_OwnTracksTests_objc_geohash : NSObject
+@end
+@implementation PodsDummy_Pods_OwnTracksTests_objc_geohash
+@end