+++ OwnTracks/Pods/Target Support Files/Pods-OwnTracksTests/Pods-OwnTracksTests-environment.h
@@ -40,11 +40,17 @@
 #define COCOAPODS_POD_AVAILABLE_MQTTClient
 #define COCOAPODS_VERSION_MAJOR_MQTTClient 0
 #define COCOAPODS_VERSION_MINOR_MQTTClient 2
-#define COCOAPODS_VERSION_PATCH_MQTTClient 0
+#define COCOAPODS_VERSION_PATCH_MQTTClient 1
 
 // QRCodeReaderViewController
 #define COCOAPODS_POD_AVAILABLE_QRCodeReaderViewController
 #define COCOAPODS_VERSION_MAJOR_QRCodeReaderViewController 3
 #define COCOAPODS_VERSION_MINOR_QRCodeReaderViewController 4
 #define COCOAPODS_VERSION_PATCH_QRCodeReaderViewController 0
+
+// objc-geohash
+#define COCOAPODS_POD_AVAILABLE_objc_geohash
+#define COCOAPODS_VERSION_MAJOR_objc_geohash 0
+#define COCOAPODS_VERSION_MINOR_objc_geohash 0
+#define COCOAPODS_VERSION_PATCH_objc_geohash 1